+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * fix version url in about_page.dart
 * guest mode fixes for media
 * fix push notifications on mastodon
+* fix async_dialog.dart: don't crash when performing async operation without actual dialog
   
 
 ## [2.6.0] [internal] 2021 July 10